extracted rules-engine code to app/utils/rules-engine.ts
